Added systemctl dependencies
This commit fixes the bug that caused the bbb-function service to start before
the network is fully up. Although the bbb-function service includes the line
`Wants=network-online.target`, we empirically verified that it only
works if both `systemd-networkd` AND `systemd-networkd-wait-online` services
are enabled, which is not the case by default in the current image.

Since other applications (like bbbread) may also need this fix, I suggest
adding it at a lower level in the image.

This commit also removes a five-second sleep that had previously been added to the
service startup as a workaround for the connection problem. We observed
that this five-second delay was insufficient for the network to become fully operational.
If an active wait is truly needed, it should be about 15 seconds to ensure the network
is completely up, as verified empirically.